requirements.txt
streamlit==1.28.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
scikit-image==0.21.0
matplotlib==3.7.2
scipy==1.11.2
scikit-learn==1.3.0
